+++ profiles/tests/test_gql_claimable_profile_query.py
@@ -1,0 +1,86 @@
+from datetime import timedelta
+from string import Template
+
+from django.utils import timezone
+from graphql_relay.node.node import to_global_id
+
+from open_city_profile.consts import TOKEN_EXPIRED_ERROR
+
+from .factories import ClaimTokenFactory, ProfileFactory
+
+
+def test_can_query_claimable_profile_with_token(rf, user_gql_client):
+    profile = ProfileFactory(user=None, first_name="John", last_name="Doe")
+    claim_token = ClaimTokenFactory(profile=profile)
+    request = rf.post("/graphql")
+    request.user = user_gql_client.user
+
+    t = Template(
+        """
+        {
+            claimableProfile(token: "${claimToken}") {
+                id
+                firstName
+                lastName
+            }
+        }
+        """
+    )
+    query = t.substitute(claimToken=claim_token.token)
+    expected_data = {
+        "claimableProfile": {
+            "id": to_global_id(type="ProfileNode", id=profile.id),
+            "firstName": profile.first_name,
+            "lastName": profile.last_name,
+        }
+    }
+    executed = user_gql_client.execute(query, context=request)
+
+    assert "errors" not in executed
+    assert dict(executed["data"]) == expected_data
+
+
+def test_cannot_query_claimable_profile_with_user_already_attached(
+    rf, user_gql_client, profile
+):
+    claim_token = ClaimTokenFactory(profile=profile)
+    request = rf.post("/graphql")
+    request.user = user_gql_client.user
+
+    t = Template(
+        """
+        {
+            claimableProfile(token: "${claimToken}") {
+                id
+            }
+        }
+        """
+    )
+    query = t.substitute(claimToken=claim_token.token)
+    executed = user_gql_client.execute(query, context=request)
+
+    assert "errors" in executed
+
+
+def test_cannot_query_claimable_profile_with_expired_token(rf, user_gql_client):
+    profile = ProfileFactory(user=None)
+    claim_token = ClaimTokenFactory(
+        profile=profile, expires_at=timezone.now() - timedelta(days=1)
+    )
+    request = rf.post("/graphql")
+    request.user = user_gql_client.user
+
+    t = Template(
+        """
+        {
+            claimableProfile(token: "${claimToken}") {
+                id
+            }
+        }
+        """
+    )
+    query = t.substitute(claimToken=claim_token.token)
+    executed = user_gql_client.execute(query, context=request)
+
+    assert "errors" in executed
+    assert executed["errors"][0]["extensions"]["code"] == TOKEN_EXPIRED_ERROR
